# api.yaml
openapi: 3.0.2
info:
  title: DBS Programmierpraktikum - Phase 4 - API-Spezifikation
  description: >-
    In der folgenden Liste finden Sie die <b>für die Anwendergruppen
    spezifischen Grundfunktionalitäten</b>. Es müssen <b>alle folgenden
    allgemeinen Bedingungen</b> eingehalten werden: <ul>
      <li><b>Bei einer gültigen <code>GET</code>-Anfrage wird ein JSON-Array von JSON-Objekten, wobei ein JSON-Objekt eine Entität darstellt, zurückgegeben</b> und dies mit einem Code <code>200</code> belegt. Wird die <code>GET</code>-Anfrage auf einer nicht vorhandenen Ressource ausgeführt, so wird mit Code <code>404</code> geantwortet. Ist die Ressource woanders zu finden, wird dies mit dem Code <code>301</code> und dem Link im Header belegt. Eine <code>GET</code>-Anfrage auf einer Ressource, die eine Sammlung von Entitäten darstellt, kann durch Übergabe von bestimmten Parametern und dazugehörigen Werten <b>case-insensitive gefiltert</b> werden. Dabei muss jede Entität in der Sammlung allen Parametern gleichzeitig genügen. Die Sammlung wird in der Antwort in ein JSON-Array gekapselt.</li>
...
      <li><b>Bei einer gültigen <code>DELETE</code>-Anfrage wird der Code <code>204</code> zurückgegeben.</b> Ist die Ressource nicht vorhanden, wird dies mit Code <code>404</code> belegt.</li>
...
      <li><b>Antworten auf gültige <code>POST</code>- und <code>PATCH</code>-Anfragen beinhalten nichts.</b></li>
...
      <li><b>Passwörter werden der Einfachheit halber im Klartext übergeben.</b></li>
...
servers:
...
paths:
...
  /alben:
    get:
      tags:
        - Anwender
      summary: Gibt Alben zurück.
      description: >-
        Alle **Alben** werden zurückgegeben. Falls Query-Parameter mit Werten
        angegeben wurden, wird die Rückgabe noch entsprechend gefiltert.
      parameters:
        - name: titel
          in: query
          description: 'Nur Alben, deren **Titel** `titel` enthält, werden zurückgegeben.'
          schema:
            type: string
        - name: schlagwort
          in: query
          description: 'Nur Alben, denen das **Schlagwort** `schlagwort` zugeordnet ist, werden zurückgegeben.'
          schema:
            type: string
        - name: durchschnittsbewertung
          in: query
          description: 'Nur Alben, deren **Durchschnittsbewertung** >= `durchschnittsbewertung` ist, werden zurückgegeben.'
          schema:
            type: number
            format: double
        - name: erstellungsdatum
          in: query
          description: 'Nur Alben, deren **Erstellungsdatum** >= `erstellungsdatum` ist, werden zurückgegeben.'
          schema:
            type: string
            format: date
      responses:
        '200':
          description: '**OK**.'
          content:
            application/json;charset=UTF-8:
              schema:
                type: array
                items:
                  allOf:
                    - type: object
                      properties:
                        blogeintragid:
                          type: integer
                          description: '**RowID des Blogeintrags**.'
                        albumid:
                          type: integer
                          description: '**RowID des Albums**.'
                      required:
                        - blogeintragid
                        - albumid
                    - $ref: '#/components/schemas/blogeintrag'
                    - $ref: '#/components/schemas/album'
    post:
      tags:
        - Authentifizierter und als Redakteur authorisierter Anwender
      summary: Ein Album wird hinzugefügt. 
      description: >-
        Ein **Album** wird hinzugefügt. Hierbei wird das **erste zugeordnete Bild** angegeben. Ein zugeordneter **Blogeintrag** wird ebenfalls angelegt.
      security:
        - basicAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              allOf:
                - type: object
                  properties:
                    bildid:
                      type: integer
                      description: '**RowID** des Bildes.'    
                  required:
                    - bildid              
                - $ref: '#/components/schemas/album'
                - $ref: '#/components/schemas/blogeintrag'
      responses:
        '201':
          $ref: '#/components/responses/201'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'                    
...
  /bilder:
    get:
      tags:
        - Authentifizierter und als Redakteur authorisierter Anwender
      summary: Gibt Bilder zurück.
      description: >-
        Alle **Bilder** werden zurückgegeben. Falls Query-Parameter mit Werten
        angegeben wurden, wird die Rückgabe noch entsprechend gefiltert.
      security:
        - basicAuth: []
      parameters:
        - name: bezeichnung
          in: query
          description: 'Nur Bilder, deren **Bezeichnung** `bezeichnung` enthält, werden zurückgegeben.'
          schema:
            type: string
        - name: aufnahmeort
          in: query
          description: 'Nur Bilder, deren **Aufnahmeort** `aufnahmeort` enthält, werden zurückgegeben.'
          schema:
            type: string            
      responses:
        '200':
          description: '**OK**.'
          content:
            application/json;charset=UTF-8:
              schema:
                type: array
                items:
                  allOf:
                    - type: object
                      properties:
                        bildid:
                          type: integer
                          description: '**RowID des Bildes**.'
                      required:
                        - bildid
                    - $ref: '#/components/schemas/bild'
    post:
      tags:
        - Authentifizierter und als Redakteur authorisierter Anwender
      summary: Fügt ein Bild hinzu.
      description: >-
        Ein **Bild** wird mit den entsprechenden Werten hinzugefügt.
      security:
        - basicAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              allOf:
                - $ref: '#/components/schemas/bild'
      responses:
        '201':
          $ref: '#/components/responses/201'
        '400':
          $ref: '#/components/responses/400'  
  '/bilder/{bildid}':
    delete:
      tags:
        - Authentifizierter und als Redakteur authorisierter Anwender
      security:
        - basicAuth: []
      summary: Löscht ein Bild.
      description: Ein **Bild** wird entfernt.
      parameters:
        - name: bildid
          in: path
          description: '**RowID** des Bilds.'
          schema:
            type: integer
          required: true
      responses:
        '204':
          $ref: '#/components/responses/204'
        '401':
          $ref: '#/components/responses/401'
        '404':
          $ref: '#/components/responses/404'
...
  '/alben/{albenid}':
    patch:
      tags:
        - Authentifizierter und als Chefredakteur authorisierter Anwender
      security:
        - basicAuth: []
      summary: Ändert ein Album.
      description: Ein **Album** wird gemäß der Query-Parameter geändert. Das **Änderungsdatum** wird entsprechend aktualisiert.
      parameters:
        - name: albenid
          in: path
          description: '**RowID** des Albums.'
          schema:
            type: integer
          required: true
        - name: sichtbarkeit
          in: query
          description: >-
            **Sichtbarkeits-Status** des Albums. Nicht sichtbar = false /
            sichtbar = true.
          schema:
            type: boolean
        - name: text
          in: query
          description: '**Text** des Albums'
          schema:
            type: string
        - name: titel
          in: query
          description: '**Titel** des Albums'
          schema:
            type: string
      responses:
        '204':
          $ref: '#/components/responses/204'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'          
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  responses:
    '201':
      description: '**Created**.'
      headers:
        Location:
          description: Die URL der neuen Ressource. Siehe hierzu auch die Anmerkung bzgl. fiktiver Links im ersten Abschnitt.
          schema:
            type: string
    '204':
      description: '**No Content**.'
...
    '400':
      description: '**Bad Request**.'
    '401':
      description: '**Unauthorized**.'
      headers:
        WWW-Authenticate:
          schema:
            type: string
    '403':
      description: '**Forbidden**.'
    '404':
      description: '**Not Found**.'
  schemas:
...
    blogeintrag:
      type: object
      properties:
        erstellungsdatum:
          type: string
          format: date-time
          description: '**Erstellungsdatum** des Blogeintrags im Format "YYYY-MM-DD HH:MM:SS".'
        aenderungsdatum:
          type: string
          format: date-time
          description: '**Änderungsdatum** des Blogeintrags im Format "YYYY-MM-DD HH:MM:SS".'          
        text:
          type: string
          description: '**Text** des Blogeintrags.'
        titel:
          type: string
          description: '**Titel** des Blogeintrags.'
      required:
        - erstellungsdatum
        - text
        - titel
...
    bild:
      type: object
      properties:
        pfad:
          type: string
          description: '**Dateipfad** des Bildes.'
        bezeichnung:
          type: string
          description: '**Bezeichnung** des Bildes.'
        aufnahmeort:
          type: string
          description: '**Aufnahmeort** des Bildes.'
      required:
        - pfad
        - bezeichnung
        - aufnahmeort
...
    album:
      type: object
      properties:
        sichtbarkeit:
          type: boolean
          description: '**Sichtbarkeit** des Albums.'
      required:
        - sichtbarkeit